requirements.txt
pillow
watchdog
wheel
pandas
pandastable
